fix: los arrastradores del sector algoritmo ya no afectan visualmente a la seccion algoritmo
