--- src/main/java/edu/fiuba/algo3/controlador/cliqueadores/BotonBorrarSectorAlgoritmoYSectorDibujoEventHandler.java
@@ -1,4 +1,0 @@
-package edu.fiuba.algo3.controlador.cliqueadores;
-
-public class BotonBorrarSectorAlgoritmoYSectorDibujoEventHandler {
-}
